Add ArgumentParser to parse arguments
Take in the CUE file to be parsed as the first positional argument and the audio
file to be split as the second, and store them as `cuefile` and `audiofile`
respectively. Also, accept the optional argument format, specified as either
`-f` or `--format`, to set the format of the file to be split.