Add command calls and output printers
Commands will be called using `subprocess.run`. Breakpoints read from
`cuebreakpoints` will then be sent to the input of `shnsplit`, which splits the
audio file. And finally, all tag information will be read from the CUE file and
written to the split files.

All lines will be printed by the program will have a unique string (`♩♬♪ `) as a
prefix. Errors will be handled by sending the STDERR of each process to it's
STDOUT, and checking the return code of the process. If it is non-zero, the
program will exit and a message displayed.